% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlsLoop2.R
\name{nlsLoop2}
\alias{nlsLoop2}
\title{Loops through a non-linear model on many different curves to find the best
possible fit}
\usage{
nlsLoop2(model, data, id_col, tries, param_bds, r2 = c("Y", "N"),
  supp.errors = c("Y", "N"), AICc = c("Y", "N"), return_preds, control, alg,
  ...)
}
\arguments{
\item{model}{a non-linear model formula, with the response on the left of a ~ operatory and an expression involving parameters on the right.}

\item{data}{the raw data frame.}

\item{id_col}{the column name that identifies each curve that is to be
fitted in "".}

\item{tries}{number iterations to use on nls2 to find best parameters}

\item{param_bds}{upper and lower boundaries for the start parameters. If
missing these default to +/- 1e+09. Need to specified as a vector as :
c(lower bound param 1, upper bound param 1, lower bound param 2, upper bound
param 2 etc)}

\item{r2}{whether or not you want the quasi rsquared value to be returned.
This defaults to no, to include the r2 values use \code{r2 = 'Y'}.}

\item{supp.errors}{if \code{supp.errors = 'Y'}, then no error messages will be shown
from the actual nlsLM function, reducing the number of error messages
received while the model works through starting parameters from which the
model cannot converge. Advised to only be used once it is expected that
error messages in the nlsLM function are not important.}

\item{AICc}{whether or not the small sample AIC should be used. Defaults to
\code{'Y'}. Override this using \code{AICc == 'N'}. AICc should be used instead of AIC
when sample size is small in comparison to the number of estimated
parameters (Burnham & Anderson 2002 recommend its use when n / n.param < 40).}

\item{return_preds}{whether you want to return the predictions with the parameters or not. Defaults to yes. Add 'N' to change to no}

\item{control}{if specific control arguments are desired they can be specified using \code{\link[minpack.lm]{nls.lm.control}}.}

\item{alg}{the algorithm with which to scan for start parameters. See \code{\link[nls2]{nls2}} for details. Defaults to \code{plinear-random}}

\item{\dots}{extra arguments to pass to \code{\link[minpack.lm]{nlsLM}} if necessary.}
}
\value{
returns a list of class \code{nlsLoop}. Notable elements within the list are \code{$params} and \code{$predictions} that give the best fit parameters and predictions based on these parameters.
}
\description{
Fits the best possible model to each of a set of curves using non-linear
least squares regression using minpack.lm::nlsLM(). Start parameters are found using nls2
}
\note{
Useful additional arguments for \code{\link[minpack.lm]{nlsLM}} include: \code{na.action = na.omit},
\code{lower/upper = c()} where these represent upper and lower boundaries for
parameter estimates
}
\examples{
# load in data

data("Chlorella_TRC")
Chlorella_TRC_test <- Chlorella_TRC[Chlorella_TRC$curve_id \%in\% c(1:10),]

# run nlsLoop()

fits <- nlsLoop2(ln.rate ~ schoolfield.high(ln.c, Ea, Eh, Th, temp = K, Tc = 20),
                data = Chlorella_TRC_test,
                tries = 500,
                id_col = 'curve_id',
                param_bds = c(-10, 10, 0.1, 2, 0.5, 5, 285, 330),
                lower = c(ln.c=-10, Ea=0, Eh=0, Th=0))

}
\seealso{
\code{\link[nlsLoop]{quasi.rsq.nls}} for details on the calculation of r squared values for non linear models.

\code{\link[minpack.lm]{nlsLM}} for details on additional arguments to pass to the nlsLM function.

\code{\link[MuMIn]{AICc}} for application of AICc.
}
\author{
Daniel Padfield
}
