% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlsViewer.R
\name{nlsViewer}
\alias{nlsViewer}
\title{an interactive addin to help evaluate the fits of nlsLoop}
\usage{
nlsViewer(data, predictions = NULL, id_col = NULL, x, y, col = NULL,
  lm_fit = FALSE)
}
\arguments{
\item{data}{the data frame that contains your raw data}

\item{predictions}{the data frame that contains your predictions if desired}

\item{id_col}{the column that splits your data frame by}

\item{x}{the x variable}

\item{y}{the y variable}

\item{col}{an optional variable if different colours are desired for each plot}

\item{lm_fit}{whether or not you want a linear model fit to be superimposed over the data. Defaults to FALSE}
}
\value{
a dataframe of the rows that are to be deleted
}
\description{
opens a pane from which you can select each set of data and select points to be dropped. The undo button gets rid of the last selection. Press "DONE" to get a dataframe of the selected outliers.
}
\details{
an interactive addin to help evaluate the fits of nlsLoop
}
\examples{
# load in data

data("Chlorella_TRC")
Chlorella_TRC_test <- Chlorella_TRC[Chlorella_TRC$curve_id \%in\% c(1:10),]

# run nlsLoop()

fits <- nlsLoop(ln.rate ~ schoolfield.high(ln.c, Ea, Eh, Th, temp = K, Tc = 20),
                data = Chlorella_TRC_test,
                tries = 500,
                id_col = 'curve_id',
                param_bds = c(-10, 10, 0.1, 2, 0.5, 5, 285, 330),
                lower = c(ln.c=-10, Ea=0, Eh=0, Th=0))

\dontrun{
outliers <- nlsViewer(data = Chlorella_TRC_test, predictions = fits$predictions,
id_col = 'curve_id', x = 'K', y = 'ln.rate')
}

}

