% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_all_nlsLoop.R
\name{plot_all_nlsLoop}
\alias{plot_all_nlsLoop}
\title{plot_all_nlsLoop}
\usage{
plot_all_nlsLoop(file_name, raw_data, param_data)
}
\arguments{
\item{file_name}{The desired path for the pdf to be saved to}

\item{raw_data}{The data frame containing the data used in the nlsLoop argument}

\item{param_data}{The nlsLoop object}
}
\value{
a plot using ggplot2 of the raw points and predictions of a single level of \code{id_col}
}
\description{
plot a pdf of all curves fitted with nlsLoop
}
\details{
Creates a pdf where each new page is a graph of one level of id_col with raw data and corresponding predictions
}
\author{
Daniel Padfield
}

