% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_curve.R
\name{plot_curve}
\alias{plot_curve}
\title{Creates a graph of one of the levels of id_col with the corresponding predictions fitted}
\usage{
plot_curve(data, predictions, id_col, id, x, y, predict.y)
}
\arguments{
\item{data}{the raw data frame}

\item{predictions}{the predictions data frame}

\item{id_col}{column name of looping factor in raw data frame}

\item{id}{the level of id_col that you want to plot}

\item{x}{the value of the independent variable}

\item{y}{the value of the dependent variable in the raw data frame}

\item{new.y}{the value of the dependent variable in the predictions data frame}
}
\value{
a plot using ggplot2 of the raw points and predictions of a single level of \code{id_col}
}
\description{
Creates a graph of one of the levels of id_col with the corresponding predictions fitted
}
\note{
variables \code{id_col}, \code{id},\code{x},\code{y}, and \code{predict.y} need to be in \code{''} or \code{""}
}
\author{
Daniel Padfield
}

