% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schoolfield.high.R
\name{schoolfield.high}
\alias{schoolfield.high}
\title{Performs the modified Sharpe-Schoolfield equation for high temperature
inactivation}
\usage{
schoolfield.high(ln.c, Ea, Eh, Th, temp, Tc)
}
\arguments{
\item{ln.c}{The specific rate of metabolism (log scale)}

\item{Ea}{activation energy (eV)}

\item{Eh}{high temperature de-activation energy (eV)}

\item{Th}{temperature at which enzyme is 1/2 active and 1/2 High-T suppressed}

\item{temp}{temperature in K}

\item{Tc}{standardisation temperature (in ºC, arbitrary)}
}
\description{
Models the thermal performance curve for metabolism using a modified thermal
performance curve.
}
\author{
Daniel Padfield
}
\references{
Schoolfield, R. M., Sharpe, P. J. & Magnuson, C. E. Non-linear regression of biological temperature-dependent rate models based on absolute reaction-rate theory. J. Theor. Biol. 88, 719–731 (1981)
}

