% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_all_nlsLoop.R
\name{plot_all_nlsLoop}
\alias{plot_all_nlsLoop}
\title{plot_all_nlsLoop}
\usage{
plot_all_nlsLoop(file_name, raw_data, param_data)
}
\arguments{
\item{file_name}{the desired path for the pdf to be saved to.}

\item{raw_data}{the data frame containing the data used in the nlsLoop argument}

\item{param_data}{the nlsLoop object}
}
\value{
a plot using ggplot2 of the raw points and predictions of a single level of \code{id_col}
}
\description{
plot a pdf of all curves fitted with nlsLoop
}
\details{
Creates a pdf where each new page is a graph of one level of \code{id_col} with raw data and corresponding predictions.
}
\examples{
data("Chlorella_TRC")
fits <- nlsLoop(ln.rate ~ schoolfield.high(ln.c, Ea, Eh, Th, temp = K, Tc = 20),
                data = Chlorella_TRC,
                tries = 500,
                id_col = 'curve_id',
                param_bds = c(-10, 10, 0.1, 2, 0.5, 5, 285, 330),
                lower = c(ln.c=-10, Ea=0, Eh=0, Th=0))

\dontrun{

plot_all_nlsLoop('WhereYouWillSaveTheFile.pdf', raw_data = Chlorella_TRC, param_data = fits)
}
}
\author{
Daniel Padfield
}

