% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_id_nlsLoop.R
\name{plot_id_nlsLoop}
\alias{plot_id_nlsLoop}
\title{Creates a graph of one of the levels of id_col with raw data and corresponding predictions}
\usage{
plot_id_nlsLoop(raw_data, param_data, id)
}
\arguments{
\item{raw_data}{The data frame containing the data used in the nlsLoop argument}

\item{param_data}{The nlsLoop object}

\item{id}{The level of id_col from which the plot is desired}
}
\value{
a plot using ggplot2 of the raw points and predictions of a single level of \code{id_col}
}
\description{
Creates a graph of one of the levels of id_col with raw data and corresponding predictions
}
\author{
Daniel Padfield
}

