% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quasi.rsq.nls.R
\name{quasi.rsq.nls}
\alias{quasi.rsq.nls}
\title{R squared statistic for non linear models}
\usage{
quasi.rsq.nls(mdl, y, param)
}
\arguments{
\item{mdl}{model object from which R^2 is desired. Used to calculate the
residual sum of squares.}

\item{y}{a vector of the response variable from the model.}

\item{param}{the number of parameters of the non-linear model}
}
\value{
value of r squared for specific non-linear model object
}
\description{
Calculates a quasi-rsquared value for non linear least squares regression
models. The formula for calculating this is given in the paper in the
references (Spiess 2010).
}
\details{
cautionary notes of whether its useful are provided in \href{http://stackoverflow.com/questions/14530770/calculating-r2-for-a-nonlinear-model}{this} StackOverflow discussion reference. A personal highlight of which is 'If someone asks for rope to hang themselves, its fine to give it to them (while performing due diligence by asking "are you sure you want to do this?")'

Calculated as :

1 - (n-1)/(n-param) * (1 - R^2)

where R^2 = 1 - RSS/TSS

RSS is the residual sum of squares given by deviance(mdl)

TSS is given by how much variance there is in the dependent variable:

(n - 1)*var(y)
}
\examples{
data(Chlorella_TRC)
fit <- minpack.lm::nlsLM(ln.rate ~ schoolfield.high(ln.c, Ea, Eh, Th, temp = K, Tc = 20),
                data = Chlorella_TRC[Chlorella_TRC$curve_id == 1,],
                start = c(ln.c = -1.2, Ea = 0.95, Eh = 5, Th = 315))

quasi.rsq.nls(fit, Chlorella_TRC[Chlorella_TRC$curve_id == 1,]$ln.rate, 4)

0.4608054
}
\references{
Spiess, A.-N. & Neumeyer, N. (2010) An evaluation of R2 as an
inadequate measure for nonlinear models in pharmacological and biochemical
research: a Monte Carlo approach. BMC pharmacology, 10, 6.

\href{http://stackoverflow.com/questions/14530770/calculating-r2-for-a-nonlinear-model}{A StackOverflow discussion}
}
\author{
Daniel Padfield
}
