% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schoolfield.high.R
\name{schoolfield.high}
\alias{schoolfield.high}
\title{Performs the modified Sharpe-Schoolfield equation for high temperature
inactivation}
\usage{
schoolfield.high(ln.c, Ea, Eh, Th, temp, Tc)
}
\arguments{
\item{ln.c}{the specific rate of metabolism (log scale).}

\item{Ea}{activation energy (eV)}

\item{Eh}{high temperature de-activation energy (eV)}

\item{Th}{temperature at which enzyme is 1/2 active and 1/2 suppressed due to high temperatures}

\item{temp}{temperature in K}

\item{Tc}{standardisation temperature (in degrees centigrade, arbitrary)}
}
\description{
Models the thermal performance curve for metabolism using a modified thermal
performance curve.
}
\examples{
temp <- seq(273.15, 333.15, 1)

plot(schoolfield.high(ln.c = -1, Ea = 0.6, Eh = 2, Th = 320, temp = temp, Tc = 20))

}
\references{
Schoolfield, R. M., Sharpe, P. J. & Magnuson, C. E. Non-linear regression of biological temperature-dependent rate models based on absolute reaction-rate theory. J. Theor. Biol. 88, 719-731 (1981)
}
\seealso{
a Shiny app showing how the parameters of the schoolfield high and full schoolfield model incorporating low temperature inactivation control the shape of the curve is available online \href{https://padpadpadpad.shinyapps.io/Sharpe-Schoolfield/}{here}
}
\author{
Daniel Padfield
}
